chore: move deployContract
